+++ utils/tests/utils.go
@@ -1,7 +1,6 @@
 package tests
 
 import (
-	"github.com/jfrog/jfrog-cli-core/v2/utils/config"
 	"github.com/jfrog/jfrog-cli-core/v2/utils/coreutils"
 	"github.com/jfrog/jfrog-client-go/utils/io/fileutils"
 	"github.com/jfrog/jfrog-client-go/utils/log"
@@ -85,43 +84,4 @@ func cleanUpUnitTestsJfrogHome(homeDir string) {
 	if errorOccurred {
 		os.Exit(1)
 	}
-}
-
-func CreateDummyJfrogConfig() (err error, cleanUp func()) {
-	err, cleanUp = SetJfrogHome()
-	if err != nil {
-		return
-	}
-	configuration := `
-		{
-		  "artifactory": [
-			{
-			  "url": "http://localhost:8080/artifactory/",
-			  "user": "user",
-			  "password": "password",
-			  "serverId": "name",
-			  "isDefault": true
-			},
-			{
-			  "url": "http://localhost:8080/artifactory/",
-			  "user": "user",
-			  "password": "password",
-			  "serverId": "notDefault"
-			}
-		  ],
-		  "version": "2"
-		}
-	`
-	content, err := config.ConvertIfNeeded([]byte(configuration))
-	if err != nil {
-		return
-	}
-	configFilePath, err := config.GetConfFilePath()
-	configFile, err := os.Create(configFilePath)
-	defer configFile.Close()
-	_, err = configFile.Write(content)
-	if err != nil {
-		return
-	}
-	return
 }
